fix(venue): resolve ui issues
